Comments
This is a combination of 7 commits. These are comments from Chris on the Pull Request. If the changes were minor, they were fixed up, and if they were more involved, they were just copied over. We did this so that we would have them in case they were to disappear during the rebase.

dataframe and groupby from pandas to handle grouping

comment addressed: "this is the kind of operation that pandas.groupby is probably much more efficient at then repeatedly testing for g == elem."

groupable attributes and type conversions

comment: "I don't like the idea of the scatter plot adding data, since it doesn't make sense in the context of other data viewers. What is the intention of the None component? Is there another way to get this to work?"

The purpose of this None component is to populate the widget with a "None" option by default, so that the scatter plot displays information normally. This was the only way I could figure out to trick the GUI into showing the option. Do you have any suggestions about where I could look to avoid adding the component?

comment: "This is a lot of type conversions, are they all necessary?"

Yeah, I cringe a bit looking at this code. I tried for a long time to get a more elegant solution, but it seems that the datetime format that num2date returns doesn't play nice and I had to typecast it to dt64. It's certainly something I can return to and give more thought.

util.py isinstance

comment: "Maybe this test could be if isinstance(data[0], (np.datetime64, datetime.date))"

Thanks! I simplified it. However, the string compare part of the test I left because when isinstance encounters datetime64 objects it has an annoying habit of displaying a unit field right after, ex. 'datetime64[ns]'

datetime property

comment: "why is the second test needed here?"

It's not! I think I was going in another direction when I made this. I took it out.

regex

comment: "quite the regex! Did you build this yourself, or pull it from somewhere else?"

I built it and it took forever. I have a weird pride in it. I hope it does all it needs to. If you see a potentially problematic gui field behavior, I can fix it up.

comments:

"There are a few places in the codebase like this now, where we have 3 code paths (one for datetimes, one for categories, one for others). This is a clue that we should probably be pulling this logic into methods on Component, which each component subtype overrides. @aak65 this isn't something that you necessarily need to worry about, but @astrofrog and I should think about it."

"Can you add a test that applies a PolygonalROI with one normal component and one datetime component, and makes sure the subset.to_mask() method works? I get an error when I make such a scatter plot and trace a shape:

  File "/Users/beaumont/glue/glue/core/subset.py", line 436, in to_mask
    result = self.roi.contains(x, y)
  File "/Users/beaumont/glue/glue/core/roi.py", line 508, in contains
    xypts = np.column_stack((x.flat, y.flat))
  File "/Users/beaumont/anaconda/lib/python2.7/site-packages/numpy/lib/shape_base.py", line 317, in column_stack
    return _nx.concatenate(arrays, 1)
TypeError: invalid type promotion
"

I haven't addressed these concerns just yet. Just creating the commit before we rebase, so we don't lose your comments.

ending message:

"Cool! I really like the datetime stuff, and I think this will be very useful, especially to non-astronomers (we have a ton of terrible unit systems, but we did the right thing with time. We just use seconds elapsed since some reference point like 400 years ago, and don't have to worry about dates, months, etc. It's awesome). I had trouble when making ROIs on scatter plots with 1 or 2 datetime axes. I also couldn't get the points to show up on a plot with one datetime axis and one non-datetime axis. How much work do you think it would be to make the histogram viewer work with datetime components? Right now they are hidden from the component dropdown. I like the groupby functionality, but I'm hesitant to pull it into the main scatter widget. Groupby is a very generically useful feature that I kind of want to see incorporated glue more fundamentally (ie something like the facetgrid functionality of tableau or seaborn, where you can turn any data viewer into a grid of plots based on a categorical grouping). What do you think about putting the groupby feature into a customized subclass of a scatterplot (maybe called GroupedScatterPlot?) Then you could still use it and we could distribute the functionality to others, perhaps as an optional add-on viewer that you enable from your config.py file."