Ensure unique WCS per Data element
Only if an extension has no WCS defined will
the Generic Fits loader combine extensions into
the same Data container.

Otherwise, the user is prompted, through the
standard glue framework, to combine like-shaped
data. Hence the user has the option of preserving
wcs information or not.

The naming scheme is also changed from shape-based
to extension-based. Whether this is a positive
change or not will remain to be seen.