Improvements to widget_properties.
CurrentComboProperty has been renamed to CurrentComboDataProperty and CurrentComboTextProperty has been added, to differentiate between cases where the item should be selected based on the text vs data in the combo box. In addition, this adds the ability to define a mapping function for value-based properties. Finally, this adds CurrentTabProperty. The tests have been expanded to cover all widget properties.