Change how we deal with component IDs for linked components.
Previously, when linking two components, we added the component ID of the linked component in each dataset to the other dataset, and set .hidden to True. However since the component ID was shared, this also hid the component ID in the original data. Instead, we now don't hide the component IDs, but instead treat component IDs where .parent is not the dataset being looked at to be effectively hidden.