Fix issues with GlueLogger
When resetting a session, sys.stderr was already set to GlueLogger and we then did self.stderr = sys.stderr and sys.stderr = self so we ended up with two references to GlueLogger. This caused issues because the original logger would try and write to the button that was in the first (now deleted) application.