Add global widget font preference #1734
This commit adds a font size feature that affects all Qt widgets' font size.
The default font size (8) has been preserved. The option is listed on the main
preferences page and uses the broadcast feature to update the UI
immediately on configuration save. This does _not_ affect the font size of
viewers or their axes.